feat: add unlock key store functionality with tests
- Implemented loadUnlockKeyStore and parseVersionedKeyStore functions to manage unlock keys.
- Added tests for loading keys from environment values and validating configurations.
- Included error handling for missing keys and malformed entries.
