+++ .env.example
@@ -19,11 +19,15 @@ PUBLIC_UNLOCK_KEY_VERSION=v1
 PUBLIC_CHAT_API_BASE=https://secret-ai-gateway.onrender.com
 
 # Serverless unlock runtime
+# Local development: use a local .env file for testing only. In staging/production,
+# store these secrets in a managed secret service such as Vercel secrets, AWS Secrets Manager,
+# GCP Secret Manager, Azure Key Vault, or another KMS-backed store.
 CHALLENGE_TOKEN_SECRET=replace-with-a-long-random-secret
 CHALLENGE_TOKEN_PREVIOUS_SECRETS=
 UNLOCK_KEY_VERSION=v1
 UNLOCK_PUBLIC_KEY=BASE64_PUBLIC_KEY
 UNLOCK_PRIVATE_KEY=BASE64_PRIVATE_KEY
-# For key rotation, optional versioned store values:
-# UNLOCK_PUBLIC_KEYS=v1:BASE64_PUBLIC_KEY,v2:BASE64_PUBLIC_KEY
-# UNLOCK_PRIVATE_KEYS=v1:BASE64_PRIVATE_KEY,v2:BASE64_PRIVATE_KEY
+# For key rotation, use versioned unlock key stores instead of a single key pair.
+# Example: UNLOCK_PUBLIC_KEYS=v1:BASE64_PUBLIC_KEY,v2:BASE64_PUBLIC_KEY
+# Example: UNLOCK_PRIVATE_KEYS=v1:BASE64_PRIVATE_KEY,v2:BASE64_PRIVATE_KEY
+# The browser should use PUBLIC_UNLOCK_PUBLIC_KEY and PUBLIC_UNLOCK_KEY_VERSION for new listings.
